+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-All notable changes to the "i18n-json-editor" extension will be documented in this file.
+All notable changes to the "ei18n-json-editor" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
