+++ media/ei18n_logo
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="utf-8"?>
+<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 500 500" xmlns:bx="https://boxy-svg.com" width="500px" height="500px"><defs/><rect x="25.985" y="25.985" width="363.253" height="363.253" rx="30" ry="30" style="fill: rgb(255, 255, 255); stroke: rgb(0, 0, 0); transform-box: fill-box; transform-origin: 50% 50%;" transform="matrix(0.707107, -0.707107, 0.707107, 0.707107, 42.388524, 42.388382)"/><g transform="matrix(1, 0, 0, 1, -0.000008, 26)"><text style="fill: rgb(51, 51, 51); font-family: Arial, sans-serif; font-size: 28px; font-weight: 700; text-anchor: middle; white-space: pre;" transform="matrix(2.413398, 0, 0, 2.413398, 89.723785, -63.645298)"><tspan x="66.411" y="111.75">Enhanced</tspan><tspan x="66.411" dy="1em">​</tspan><tspan x="66.411" dy="1em">​</tspan><tspan>Editor</tspan></text><g transform="matrix(1, 0, 0, 1, 1.126671, -0.000015)"><text style="fill: rgb(204, 88, 3); font-family: Arial, sans-serif; font-size: 28px; font-weight: 700; text-anchor: middle; white-space: pre;" transform="matrix(2.413398, 0, 0, 2.413398, 9.979706, -63.64529)"><tspan x="66.411" y="111.75" dy="1em">​</tspan><tspan>i18n</tspan></text><text style="fill: rgb(29, 95, 165); font-family: Arial, sans-serif; font-size: 28px; font-weight: 700; text-anchor: middle; white-space: pre;" transform="matrix(2.413398, 0, 0, 2.413398, 155.979706, -63.64529)"><tspan x="66.411" y="111.75" dy="1em">​</tspan><tspan>.json</tspan></text></g></g></svg>
