+++ media/template.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     {{LINKS}}
     {{SCRIPTS}}
-    <title>i18n-json-editor</title>
+    <title>ei18n-json-editor</title>
 </head>
 
 <body>
